+++ src/builds/mewcx/cxHelpers/background.js
@@ -30,9 +30,11 @@ const networkChanger = items => {
     const networkProps = JSON.parse(items['defNetwork']);
     let network = {};
     if (networkProps.hasOwnProperty('url')) {
-      network = store.state.Networks[networkProps.key].find(actualNetwork => {
-        return actualNetwork.url === networkProps.url;
-      });
+      network = store.state.main.Networks[networkProps.key].find(
+        actualNetwork => {
+          return actualNetwork.url === networkProps.url;
+        }
+      );
 
       chrome.storage.sync.set({
         defNetwork: JSON.stringify({
@@ -41,17 +43,19 @@ const networkChanger = items => {
         })
       });
     } else {
-      network = store.state.Networks[networkProps.key].find(actualNetwork => {
-        return actualNetwork.service === networkProps.service;
-      });
+      network = store.state.main.Networks[networkProps.key].find(
+        actualNetwork => {
+          return actualNetwork.service === networkProps.service;
+        }
+      );
     }
     // eslint-disable-next-line
     if (!!network) {
-      store.dispatch('switchNetwork', network).then(() => {
-        store.dispatch('setWeb3Instance', network.url).then(() => {
-          store.state.web3.eth.net.getId().then(res => {
+      store.dispatch('main/switchNetwork', network).then(() => {
+        store.dispatch('main/setWeb3Instance', network.url).then(() => {
+          store.state.main.web3.eth.net.getId().then(res => {
             chrome.storage.sync.set({
-              defChainID: store.state.network.type.chainID,
+              defChainID: store.state.main.network.type.chainID,
               defNetVersion: res
             });
           });
@@ -59,14 +63,14 @@ const networkChanger = items => {
       });
     }
   } else {
-    store.dispatch('setWeb3Instance');
-    store.state.web3.eth.net.getId().then(res => {
+    store.dispatch('main/setWeb3Instance');
+    store.state.main.web3.eth.net.getId().then(res => {
       chrome.storage.sync.set({
-        defChainID: store.state.network.type.chainID,
+        defChainID: store.state.main.network.type.chainID,
         defNetVersion: res,
         defNetwork: JSON.stringify({
-          service: store.state.network.service,
-          key: store.state.network.type.name
+          service: store.state.main.network.service,
+          key: store.state.main.network.type.name
         })
       });
     });
@@ -92,21 +96,21 @@ chrome.storage.onChanged.addListener(items => {
       const networkProps = JSON.parse(
         Misc.stripTags(items['defNetwork'].newValue)
       );
-      const network = store.state.Networks[networkProps.key].find(
+      const network = store.state.main.Networks[networkProps.key].find(
         actualNetwork => {
           return actualNetwork.service === networkProps.service;
         }
       );
       store
         .dispatch(
-          'switchNetwork',
-          network ? store.state.Networks[networkProps.key][0] : network
+          'main/switchNetwork',
+          network ? store.state.main.Networks[networkProps.key][0] : network
         )
         .then(() => {
-          store.dispatch('setWeb3Instance', network.url).then(() => {
-            store.state.web3.eth.net.getId().then(res => {
+          store.dispatch('main/setWeb3Instance', network.url).then(() => {
+            store.state.main.web3.eth.net.getId().then(res => {
               chrome.storage.sync.set({
-                defChainID: store.state.network.type.chainID,
+                defChainID: store.state.main.network.type.chainID,
                 defNetVersion: res
               });
             });
@@ -249,7 +253,7 @@ function querycB(tab) {
       } else {
         // Injects web3
         chrome.tabs.sendMessage(tab.id, { event: CX_INJECT_WEB3 }, function() {
-          store.state.web3.eth.net.getId().then(() => {
+          store.state.main.web3.eth.net.getId().then(() => {
             chrome.tabs.sendMessage(tab.id, {
               event: WEB3_INJECT_SUCCESS.replace('{{id}}', 'internal') // triggers connect call
             });
@@ -259,7 +263,7 @@ function querycB(tab) {
     } else {
       // Injects web3
       chrome.tabs.sendMessage(tab.id, { event: CX_INJECT_WEB3 }, function() {
-        store.state.web3.eth.net.getId().then(() => {
+        store.state.main.web3.eth.net.getId().then(() => {
           chrome.tabs.sendMessage(tab.id, {
             event: WEB3_INJECT_SUCCESS.replace('{{id}}', 'internal') // triggers connect call
           });