+++ src/builds/web/app.vue
@@ -20,7 +20,7 @@ import HeaderContainer from '@/containers/HeaderContainer';
 import ConfirmationContainer from '@/containers/ConfirmationContainer';
 import WelcomeModal from '@/components/WelcomeModal';
 import store from 'store';
-import { mapState } from 'vuex';
+import { mapState, mapActions } from 'vuex';
 import { Toast } from '@/helpers';
 import LogoutWarningModal from '@/components/LogoutWarningModal';
 
@@ -34,7 +34,7 @@ export default {
     'welcome-modal': WelcomeModal
   },
   computed: {
-    ...mapState(['wallet', 'online'])
+    ...mapState('main', ['wallet', 'online'])
   },
   watch: {
     $route(to, from) {
@@ -57,14 +57,14 @@ export default {
       Toast.responseHandler(errMsg, Toast.WARN);
     });
     window.addEventListener('online', () => {
-      this.$store.dispatch('checkIfOnline', true);
+      this.checkIfOnline(true);
     });
     window.addEventListener('offline', () => {
-      this.$store.dispatch('checkIfOnline', false);
+      this.checkIfOnline(false);
     });
   },
   mounted() {
-    this.$store.dispatch('checkIfOnline', navigator.onLine);
+    this.checkIfOnline(navigator.onLine);
     if (!store.get('notFirstTimeVisit') && this.$route.fullPath === '/') {
       this.$refs.welcome.$refs.welcome.show();
     }
@@ -81,6 +81,9 @@ export default {
     window.removeEventListener('PWA_UPDATED');
     window.removeEventListener('offline');
     window.removeEventListener('online');
+  },
+  methods: {
+    ...mapActions('main', ['checkIfOnline'])
   }
 };
 </script>