+++ src/containers/NotificationsContainer/components/NotificationTypes/NotificationDetails/TransactionDetails.vue
@@ -155,7 +155,7 @@ export default {
     };
   },
   computed: {
-    ...mapState(['web3', 'network', 'notifications', 'wallet']),
+    ...mapState('main', ['web3', 'network', 'notifications', 'wallet']),
     errorMessage() {
       return this.errorMessageString(this.notice);
     },